# openapi.yaml
openapi: 3.1.0
info:
  title: Hub Innowacji Spolecznych API
  description: |-
    REST API dla mieszkańców, pracowników i administratora.
    Publiczne: katalog, regiony, zatwierdzone pomysły, dopasowanie i asystent.
    Konto mieszkańca jest wymagane do zgłoszenia nowego pomysłu i daje dostęp
    do /api/v1/ideas?mine=true. /api/v1/me oraz raporty pracownicze wymagają
    sesji; /api/v1/admin/** wymaga roli ADMIN, a /api/v1/staff/**
    roli STAFF lub ADMIN. Zgłoszenie pomysłu powiadamia pracowników;
    zalogowany autor może odczytać odpowiedzi przez /api/v1/ideas/{id}/replies
    oraz /api/v1/notifications. Trendy są w /api/v1/admin/trends;
    administrator może edytować i publikować wiedzę.
    Rejestracja: POST /api/v1/register. Logowanie: POST /login jako formularz
    username (adres e-mail), password, _csrf; odpowiedź ustawia JSESSIONID.
    Przed POST/PUT/PATCH pobierz GET /api/v1/csrf i wyślij token w nagłówku
    X-CSRF-TOKEN. Po logowaniu pobierz token ponownie. W fetch użyj
    credentials: 'include'.
  version: v1
servers:
- url: http://localhost
  description: Generated server url
paths:
...
  /api/v1/admin/innovations/{id}:
    get:
      tags:
      - knowledge-admin-controller
      operationId: getInnovation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InnovationDto"
      security:
      - cookieAuth: []
    put:
      tags:
      - knowledge-admin-controller
      operationId: updateInnovation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: X-CSRF-TOKEN
        in: header
        required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InnovationUpdate"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InnovationDto"
      security:
      - cookieAuth: []
...
  /api/v1/matches:
    post:
      tags:
      - api-controller
      operationId: match
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchResponse"
      security:
      - csrfToken: []
  /api/v1/ideas:
    get:
      tags:
      - api-controller
      operationId: listIdeas
      parameters:
      - name: mine
        in: query
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IdeaDto"
    post:
      tags:
      - api-controller
      operationId: createIdea
      parameters:
      - name: X-CSRF-TOKEN
        in: header
        required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdeaRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IdeaDto"
      security:
      - cookieAuth: []
  /api/v1/ideas/assistant:
    post:
      tags:
      - api-controller
      operationId: assistant
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssistantReply"
      security:
      - csrfToken: []
  /api/v1/admin/resources:
    get:
      tags:
      - knowledge-admin-controller
      operationId: listResources
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResourceDto"
      security:
      - cookieAuth: []
    post:
      tags:
      - knowledge-admin-controller
      operationId: createResource
      parameters:
      - name: X-CSRF-TOKEN
        in: header
        required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceUpdate"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResourceDto"
      security:
      - cookieAuth: []
  /api/v1/admin/innovations:
    get:
      tags:
      - knowledge-admin-controller
      operationId: listInnovations
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageDtoInnovationDto"
      security:
      - cookieAuth: []
    post:
      tags:
      - api-controller
      operationId: addInnovation
      parameters:
      - name: X-CSRF-TOKEN
        in: header
        required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InnovationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InnovationDto"
      security:
      - cookieAuth: []
...
  /api/v1/admin/reports/{id}/status:
    patch:
      tags:
      - api-controller
      operationId: setReportStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: X-CSRF-TOKEN
        in: header
        required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReportDto"
      security:
      - cookieAuth: []
  /api/v1/admin/ideas/{id}/moderation:
    patch:
      tags:
      - api-controller
      operationId: moderateIdea
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: X-CSRF-TOKEN
        in: header
        required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModerationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IdeaDto"
      security:
      - cookieAuth: []
...
  /api/v1/reports:
    get:
      tags:
      - api-controller
      operationId: listReports
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportDto"
      security:
      - cookieAuth: []
...
  /api/v1/me:
    get:
      tags:
      - api-controller
      operationId: me
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MeDto"
      security:
      - cookieAuth: []
  /api/v1/matches/{reportId}:
    get:
      tags:
      - api-controller
      operationId: getMatch
      parameters:
      - name: reportId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchResponse"
  /api/v1/innovations:
    get:
      tags:
      - api-controller
      operationId: listInnovations_1
      parameters:
      - name: areaId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: region
        in: query
        required: false
        schema:
          type: string
          enum:
          - DOLNOSLASKIE
          - KUJAWSKO_POMORSKIE
          - LUBELSKIE
          - LUBUSKIE
          - LODZKIE
          - MALOPOLSKA
          - MAZOWIECKIE
          - OPOLSKIE
          - PODKARPACKIE
          - PODLASKIE
          - POMORSKIE
          - SLASKIE
          - SWIETOKRZYSKIE
          - WARMINSKO_MAZURSKIE
          - WIELKOPOLSKIE
          - ZACHODNIOPOMORSKIE
          - POLSKA
          - INNE
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: "createdAt,desc"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageDtoInnovationDto"
  /api/v1/innovations/{id}:
    get:
      tags:
      - api-controller
      operationId: getInnovation_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InnovationDto"
  /api/v1/ideas/{id}:
    get:
      tags:
      - api-controller
      operationId: getIdea
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IdeaDto"
...
  /api/v1/admin/ideas:
    get:
      tags:
      - api-controller
      operationId: listIdeasForModeration
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          default: PENDING
          enum:
          - PENDING
          - APPROVED
          - REJECTED
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IdeaDto"
      security:
      - cookieAuth: []
components:
  schemas:
...
    InnovationUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 0
        summary:
          type: string
          maxLength: 1000
          minLength: 0
        description:
          type: string
        targetGroup:
          type: string
          maxLength: 255
          minLength: 0
        status:
          type: string
          enum:
          - ROZWOJ
          - TESTOWANA
          - WDROZONA
        region:
          type: string
          enum:
          - DOLNOSLASKIE
          - KUJAWSKO_POMORSKIE
          - LUBELSKIE
          - LUBUSKIE
          - LODZKIE
          - MALOPOLSKA
          - MAZOWIECKIE
          - OPOLSKIE
          - PODKARPACKIE
          - PODLASKIE
          - POMORSKIE
          - SLASKIE
          - SWIETOKRZYSKIE
          - WARMINSKO_MAZURSKIE
          - WIELKOPOLSKIE
          - ZACHODNIOPOMORSKIE
          - POLSKA
          - INNE
        videoUrl:
          type: string
          maxLength: 500
          minLength: 0
          pattern: https?://.*
        sourceUrl:
          type: string
          maxLength: 500
          minLength: 0
          pattern: https?://.*
        areaId:
          type: integer
          format: int64
        published:
          type: boolean
      required:
      - published
      - status
      - title
...
    InnovationDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        summary:
          type: string
        description:
          type: string
        targetGroup:
          type: string
        status:
          type: string
          enum:
          - ROZWOJ
          - TESTOWANA
          - WDROZONA
        region:
          type: string
          enum:
          - DOLNOSLASKIE
          - KUJAWSKO_POMORSKIE
          - LUBELSKIE
          - LUBUSKIE
          - LODZKIE
          - MALOPOLSKA
          - MAZOWIECKIE
          - OPOLSKIE
          - PODKARPACKIE
          - PODLASKIE
          - POMORSKIE
          - SLASKIE
          - SWIETOKRZYSKIE
          - WARMINSKO_MAZURSKIE
          - WIELKOPOLSKIE
          - ZACHODNIOPOMORSKIE
          - POLSKA
          - INNE
        videoUrl:
          type: string
        sourceUrl:
          type: string
        area:
          $ref: "#/components/schemas/AreaDto"
        createdAt:
          type: string
          format: date-time
        published:
          type: boolean
...
    MatchRequest:
      type: object
      properties:
        description:
          type: string
          minLength: 1
        region:
          type: string
          enum:
          - DOLNOSLASKIE
          - KUJAWSKO_POMORSKIE
          - LUBELSKIE
          - LUBUSKIE
          - LODZKIE
          - MALOPOLSKA
          - MAZOWIECKIE
          - OPOLSKIE
          - PODKARPACKIE
          - PODLASKIE
          - POMORSKIE
          - SLASKIE
          - SWIETOKRZYSKIE
          - WARMINSKO_MAZURSKIE
          - WIELKOPOLSKIE
          - ZACHODNIOPOMORSKIE
          - POLSKA
          - INNE
        authorName:
          type: string
      required:
      - description
...
    AssistantReply:
      type: object
      properties:
        reply:
          type: string
...
    ModerationRequest:
      type: object
      properties:
        status:
          type: string
          enum:
          - PENDING
          - APPROVED
          - REJECTED
      required:
      - status
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: JSESSIONID
      in: cookie
    csrfToken:
      type: apiKey
      name: X-CSRF-TOKEN
      in: header